Add promhttp_start_daemon_with_options() function.
This is the same as promhttp_start_daemon(), except that it exposes
MHD_start_daemon()'s variadic options, to allow applications to pass flags
directly into libmicrohttpd. The use case for this functionality is to allow
the metrics listener to be bound to a specific socket address (instead of
binding to a given port on the wildcard address) using MHD_OPTION_SOCK_ADDR,
etc.