Declare promhttp_handler with the correct return type
The type of MHD_YES/NO changed in libmicrohttpd 0.9.70, so flexibility
is needed for some time to avoid "incompatible pointer type" warning,
which breaks the default build where warnings are treated as errors.

Furthermore the public declaration of promhttp_handler() is useful
when the promhttpd_start_daemon() helper is not flexible enough and
thus the library user has to invoke MHD_start_daemon() directly.

Fixes #72.